# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "wwg API",
    "description": "The Wasatch Wargamers Campaign App API.",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/masquerade/end": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Ends a masquerade (decision 0012): back to the Admin's own session, without a password, if\nthe masquerade may still go on. 409 if the session isn't a masquerade.",
        "operationId": "EndMasquerade",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/campaigns/{id}/armies": {
      "get": {
        "tags": [
          "Armies"
        ],
        "summary": "Every army in the campaign, with its commander, sorted by name. All members see them all,\nunassigned ones included.",
        "operationId": "ListArmies",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArmySummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Armies"
        ],
        "summary": "Adds an army, optionally with a commander, side and factions (Umpire or Admin). A campaign has at\nmost 8 armies (409). Without a colour it gets the first one no other army has.",
        "operationId": "CreateArmy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArmyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/armies/{id}": {
      "get": {
        "tags": [
          "Armies"
        ],
        "summary": "An army's details and units. Every member sees every army's (read-only); where the units\nare is a separate question (DESIGN.md §5.2, the visibility rule).",
        "operationId": "GetArmy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Armies"
        ],
        "summary": "Changes an army's name, side, colour, nation and factions (Umpire or Admin). A faction the\narmy has units from can't be dropped (409).",
        "operationId": "UpdateArmy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArmyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Armies"
        ],
        "summary": "Deletes an army and its units (Umpire or Admin), while the campaign is setting up; after it\nstarts, that would erase their history (409).",
        "operationId": "DeleteArmy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/armies/{id}/commander": {
      "put": {
        "tags": [
          "Armies"
        ],
        "summary": "Makes a Player the army's commander, replacing any other (Umpire or Admin). A Player\ncommands at most one army: 409 if they already command another.",
        "operationId": "AssignCommander",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCommanderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Armies"
        ],
        "summary": "Leaves the army without a commander (Umpire or Admin).",
        "operationId": "UnassignCommander",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArmyResponse": {
        "required": [
          "id",
          "campaignId",
          "campaignName",
          "name",
          "commander",
          "side",
          "color",
          "nation",
          "factions",
          "units",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The army's ID.",
            "format": "uuid"
          },
          "campaignId": {
            "type": "string",
            "description": "The campaign it's in.",
            "format": "uuid"
          },
          "campaignName": {
            "type": "string",
            "description": "That campaign's name."
          },
          "name": {
            "type": "string",
            "description": "Its name."
          },
          "commander": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Its commander, or null if unassigned.",
                "$ref": "#/components/schemas/ArmyCommander"
              }
            ]
          },
          "side": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Its side, or null if it has none yet.",
                "$ref": "#/components/schemas/ArmySide"
              }
            ]
          },
          "color": {
            "description": "Its colour.",
            "$ref": "#/components/schemas/ArmyColor"
          },
          "nation": {
            "description": "The nation it fights for, drawn as its flag.",
            "$ref": "#/components/schemas/Nation"
          },
          "factions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArmyFactionResponse"
            },
            "description": "The library factions it takes its units from, sorted by name."
          },
          "units": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArmyUnitResponse"
            },
            "description": "Its units, sorted by name."
          },
          "createdAt": {
            "type": "string",
            "description": "When it was created (UTC).",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "When it last changed (UTC).",
            "format": "date-time"
          }
        },
        "description": "An army's details, with its units."
      },
      "ArmySide": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The side's ID.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Its name."
          }
        },
        "description": "The side an army is in."
      },
      "ArmySummary": {
        "required": [
          "id",
          "name",
          "commander",
          "side",
          "color",
          "nation"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The army's ID.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Its name."
          },
          "commander": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Its commander, or null if unassigned.",
                "$ref": "#/components/schemas/ArmyCommander"
              }
            ]
          },
          "side": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Its side, or null if it has none yet.",
                "$ref": "#/components/schemas/ArmySide"
              }
            ]
          },
          "color": {
            "description": "Its colour.",
            "$ref": "#/components/schemas/ArmyColor"
          },
          "nation": {
            "description": "The nation it fights for, drawn as its flag.",
            "$ref": "#/components/schemas/Nation"
          }
        },
        "description": "An army in the campaign's army list. Every member sees every army."
      },
...
      "AssignCommanderRequest": {
        "required": [
          "memberId"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "description": "The Player's membership ID.",
            "format": "uuid"
          }
        },
        "description": "Gives an army a commander."
      },
...
      "CreateArmyRequest": {
        "required": [
          "name",
          "commanderMemberId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string",
            "description": "The army's name."
          },
          "commanderMemberId": {
            "type": [
              "null",
              "string"
            ],
            "description": "A Player to command it (their membership ID), or null.",
            "format": "uuid"
          },
          "sideId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Its side, or null for none yet (\"Unassigned\").",
            "format": "uuid"
          },
          "color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Its colour, or null for the first one no other army has.",
                "$ref": "#/components/schemas/ArmyColor"
              }
            ]
          },
          "nation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "The nation it fights for (its flag), or null for none (a plain flag).",
                "$ref": "#/components/schemas/Nation"
              }
            ]
          },
          "factionIds": {
            "maxItems": 20,
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The library factions it takes its units from, or null for none yet."
          }
        },
        "description": "Adds an army to the campaign (at most 8)."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TokenResponse": {
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The bearer token for API calls."
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds until the access token expires; refresh before then.",
            "format": "int32"
          }
        },
        "description": "A signed-in session. The access token goes in `Authorization: Bearer …`; the refresh\ntoken is never in a body: it's set as an HttpOnly cookie that only /api/auth receives."
      },
...
      "UpdateArmyRequest": {
        "required": [
          "name",
          "sideId",
          "color",
          "nation"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string",
            "description": "The army's name."
          },
          "sideId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Its side, or null for none (\"Unassigned\").",
            "format": "uuid"
          },
          "color": {
            "description": "Its colour. Two armies can share one.",
            "$ref": "#/components/schemas/ArmyColor"
          },
          "nation": {
            "description": "The nation it fights for, drawn as its flag.",
            "$ref": "#/components/schemas/Nation"
          },
          "factionIds": {
            "maxItems": 20,
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The library factions it takes its units from (one it has units from can't be dropped), or null\nto leave them as they are."
          }
        },
        "description": "Changes an army's name, side, colour, nation and factions."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "The accessToken from register, login or refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}